requirements.txt
# 核心语音识别和音频处理
openai-whisper>=20231117
SpeechRecognition>=3.10.0
pyaudio>=0.2.11
librosa>=0.10.1
soundfile>=0.12.1
numpy>=1.24.3

# GUI界面和系统集成
PyQt5>=5.15.10
pyperclip>=1.8.2

# 可视化和图表
matplotlib>=3.7.1

# 全局热键支持
keyboard>=0.13.5

# 繁简转换
opencc-python-reimplemented>=0.1.7

# 可选依赖（如需要可单独安装）
# QHotkey  # 跨平台全局热键，可选
# psutil   # 系统信息监控，可选
# plyer    # 跨平台通知，可选